backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pdfplumber==0.10.3
requests==2.31.0
jieba==0.42.1
